working filters on depth images

+++ libgesture/include/lgfreenectdevice.h
@@ -2,7 +2,7 @@
 #ifndef DARMA_LG_FREENECT_DEVICE
 #define DARMA_LG_FREENECT_DEVICE
 
-#include "libfreenect.hpp"
+#include <libfreenect.hpp>
 #include "mutex.h"
 #include <cv.h>
 #include <iostream>
@@ -23,7 +23,7 @@ class LGFreenectDevice : public Freenect::FreenectDevice
 		bool getDepth(cv::Mat& output);
 
 	private:
-		std::vector<uint8_t> m_buffer_depth;
+		std::vector<uint16_t> m_buffer_depth;
 		std::vector<uint8_t> m_buffer_rgb;
 		std::vector<uint16_t> m_gamma;
 		cv::Mat depthMat;
